각종 에러 fix 및 Lecture #7.4 finished

+++ bangtube/controllers/videoController.js
@@ -34,16 +34,20 @@ export const postUpload = async (req, res) => {
   const {
     body: { title, description },
     file: { location }
+    // user: {  id: userId  }
   } = req;
   const newVideo = await Video.create({
     fileUrl: location,
-    title: title,
-    description: description,
-    creator: req.user.id
+    title,
+    description,
+    creator: req.user._id
   });
-  console.log(req.user.videos);
   req.user.videos.push(newVideo.id);
-  // req.user.save();
+  try {
+    req.user.save();
+  } catch (error) {
+    console.log(error);
+  }
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -52,8 +56,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
-    console.log(video);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -66,7 +69,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if(video.creator !== req.user.id) {
+      throw Error();      
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -90,7 +97,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if(video.creator !== req.user.id) {
+      throw Error();      
+    } else {
+      await Video.findByIdAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
